requirements.txt
appdirs==1.4.3
Django==1.11
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyenchant==1.6.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
